Keep connector ids on codeblock after undo...

+++ src/DynamoCore/Core/UndoRedoRecorder.cs
@@ -94,7 +94,6 @@ namespace Dynamo.Core
 
             undoStack = new Stack<XmlElement>();
             redoStack = new Stack<XmlElement>();
-            offTrackModels = new HashSet<Guid>();
         }
 
         /// <summary>
@@ -173,7 +172,6 @@ namespace Dynamo.Core
             EnsureValidRecorderStates();
             undoStack.Clear();
             redoStack.Clear();
-            offTrackModels.Clear();
         }
 
         #endregion
@@ -244,24 +242,6 @@ namespace Dynamo.Core
             }
 
             return PopActionGroupFromUndoStack();
-        }
-
-        /// <summary>
-        /// A model is recorded as an off-track object means it is modified
-        /// somewhere else, but that modification operation is not recorded in
-        /// undo/redo stack. UndoRedoRecorder will ignore all excpetions that
-        /// related to this kind of objects during undo/redo.
-        /// 
-        /// For example, a connector that connects to an input port of a custom
-        /// node instance could be deleted because of the removal of that input
-        /// port in custom workspace. As this deletion in the custom workspace 
-        /// is not recorded by UndoRedoRecorder in home workspace, the connector
-        /// should be marked as off-track.
-        /// </summary>
-        /// <param name="modelGuid"></param>
-        public void RecordModelAsOffTrack(Guid modelGuid)
-        {
-            offTrackModels.Add(modelGuid);
         }
         #endregion
 
@@ -395,33 +375,23 @@ namespace Dynamo.Core
                     // up-to-date model is retrieved and serialized into the 
                     // redo action group so that it can properly be redone later.
                     case UserAction.Creation:
-                        try
+
+                        ModelBase toBeDeleted = undoClient.GetModelForElement(element);
+                        if (toBeDeleted != null)
                         {
-                            ModelBase toBeDeleted = undoClient.GetModelForElement(element);
                             RecordActionInternal(newGroup, toBeDeleted, modelActionType);
                             undoClient.DeleteModel(element);
-                        }
-                        catch (ArgumentException e)
-                        {
-                            bool isOffTrackObject = false;
-                            var guidAttribute = element.Attributes["guid"];
-                            if (guidAttribute != null)
-                            {
-                                var guid = Guid.Parse(guidAttribute.Value);
-                                isOffTrackObject = offTrackModels.Contains(guid);
-                            }
-
-                            if (!isOffTrackObject)
-                            {
-                                throw e;
-                            }
                         }
                         break;
 
                     case UserAction.Modification:
                         ModelBase toBeUpdated = undoClient.GetModelForElement(element);
-                        RecordActionInternal(newGroup, toBeUpdated, modelActionType);
-                        undoClient.ReloadModel(element);
+                        if (toBeUpdated != null)
+                        {
+
+                            RecordActionInternal(newGroup, toBeUpdated, modelActionType);
+                            undoClient.ReloadModel(element);
+                        }
                         break;
 
                     case UserAction.Deletion:
@@ -457,14 +427,21 @@ namespace Dynamo.Core
 
                     case UserAction.Modification:
                         ModelBase toBeUpdated = undoClient.GetModelForElement(element);
-                        RecordActionInternal(newGroup, toBeUpdated, modelActionType);
-                        undoClient.ReloadModel(element);
+                        if (toBeUpdated != null)
+                        {
+
+                            RecordActionInternal(newGroup, toBeUpdated, modelActionType);
+                            undoClient.ReloadModel(element);
+                        }
                         break;
 
                     case UserAction.Deletion:
                         ModelBase toBeDeleted = undoClient.GetModelForElement(element);
-                        RecordActionInternal(newGroup, toBeDeleted, modelActionType);
-                        undoClient.DeleteModel(element);
+                        if (toBeDeleted != null)
+                        {
+                            RecordActionInternal(newGroup, toBeDeleted, modelActionType);
+                            undoClient.DeleteModel(element);
+                        }
                         break;
                 }
             }
